// packages/backend/src/modules/jobs-integration/types/TTheirStackJobsResponse.ts
export type TTheirStackJobsResponse = {
  metadata: {
    total_results: number;
    truncated_results: number;
    truncated_companies: number;
    total_companies: number;
  };
  data: Array<{
    id: number;
    job_title: string;
    url: string;
    date_posted: string;
    has_blurred_data: boolean;
    company: string;
    final_url: string;
    source_url: string;
    location: string;
    short_location: string;
    long_location: string;
    state_code: string;
    latitude: number;
    longitude: number;
    postal_code: string;
    remote: boolean;
    hybrid: boolean;
    salary_string: string;
    min_annual_salary: number;
    min_annual_salary_usd: number;
    max_annual_salary: number;
    max_annual_salary_usd: number;
    avg_annual_salary_usd: number;
    salary_currency: string;
    countries: string[];
    country: string;
    cities: string[];
    continents: string[];
    seniority: string;
    country_codes: string[];
    country_code: string;
    discovered_at: string;
    hiring_team: Array<{
      first_name: string;
      full_name: string;
      linkedin_url: string;
      role: string;
      image_url: string;
      thumbnail_url: string;
    }>;
    company_domain: string;
    reposted: boolean;
    date_reposted: string;
    description: string;
    company_object: {
      name: string;
      domain: string;
      industry: string;
      country: string;
      country_code: string;
      employee_count: number;
      logo: string;
      num_jobs: number;
      num_technologies: number;
      url: string;
      industry_id: number;
      linkedin_url: string;
      num_jobs_last_30_days: number;
      num_jobs_found: number;
      yc_batch: string;
      apollo_id: string;
      linkedin_id: string;
      url_source: string;
      is_recruiting_agency: boolean;
      id: string;
      founded_year: number;
      annual_revenue_usd: number;
      annual_revenue_usd_readable: string;
      total_funding_usd: number;
      last_funding_round_date: string;
      last_funding_round_amount_readable: string;
      employee_count_range: string;
      long_description: string;
      seo_description: string;
      city: string;
      postal_code: string;
      company_keywords: string[];
      alexa_ranking: number;
      publicly_traded_symbol: string;
      publicly_traded_exchange: string;
      investors: string[];
      funding_stage: string;
      has_blurred_data: boolean;
      technology_slugs: string[];
      technology_names: string[];
    };
    normalized_title: string;
    manager_roles: string[];
    matching_phrases: string[];
    matching_words: string[];
  }>;
};